push final backend

+++ app/Http/Controllers/KamarController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Checkin;
 use App\Models\Kamar;
 use App\Models\TypeKamar;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class KamarController extends Controller
 {
@@ -18,8 +20,8 @@ class KamarController extends Controller
     {
         $type_kamars = TypeKamar::all();
         $kamars = DB::table('kamars')
-            ->join('type_kamars', 'kamars.id_type_kamar', '=', 'type_kamars.id')
-            ->select('kamars.id', 'kamars.no_kamar', 'kamars.id_type_kamar', 'type_kamars.jenis', 'kamars.max', 'kamars.status')
+            ->join('type_kamars', 'kamars.type_kamar_id', '=', 'type_kamars.id')
+            ->select('kamars.id', 'kamars.no_kamar', 'kamars.type_kamar_id', 'type_kamars.jenis', 'kamars.max', 'kamars.status')
             ->get();
         return view('kamar.index', compact('kamars', 'type_kamars'));
     }
@@ -46,12 +48,15 @@ class KamarController extends Controller
 
         $validateData = $request->validate([
             'no_kamar' => 'required',
-            'id_type_kamar' => 'required',
+            'type_kamar_id' => 'required',
             'max' => 'required',
             'status' => 'required',
         ]);
 
         Kamar::create($validateData);
+
+        Alert::success("Successfully", "Kamar " . $request->no_kamar . " tidak ditemukan!");
+
         return redirect('/kamar');
     }
 
@@ -76,8 +81,8 @@ class KamarController extends Controller
     {
         //get All data dari type kamar
         $type_kamars = TypeKamar::all();
-        //get 1 data dari type kamar dimana id type kamar = id_type_kamar dari tabel kamar
-        $type_kamar = TypeKamar::where('id', '=', $kamar->id_type_kamar)->first();
+        //get 1 data dari type kamar dimana id type kamar = type_kamar_id dari tabel kamar
+        $type_kamar = TypeKamar::where('id', '=', $kamar->type_kamar_id)->first();
         //menambahkan collection jenis ke $kamar
         $kamar->jenis = $type_kamar->jenis;
 
@@ -96,12 +101,13 @@ class KamarController extends Controller
 
         $validateData = $request->validate([
             'no_kamar' => 'required',
-            'id_type_kamar' => 'required',
+            'type_kamar_id' => 'required',
             'max' => 'required',
             'status' => 'required',
         ]);
 
         $kamar->update($validateData);
+        Alert::success("Yeay!", "Kamar " . $request->no_kamar . " berhasil diupdate!");
         return redirect('/kamar')->with('pesan', "Kamar $request->no_kamar berhasil di Update");
     }
 
@@ -113,13 +119,26 @@ class KamarController extends Controller
      */
     public function destroy(Kamar $kamar)
     {
-        $kamar->delete();
+        $checkins = Checkin::where('kamar_id', '=', $kamar->id)->count();
+
+
+        if ($checkins == 0) {
+            $kamar->delete();
+            Alert::success('Sukses', "Data kamar " . $kamar->no_kamar . " berhasil dihapus!");
+        } else {
+            Alert::error('Gagal', "Data kamar " . $kamar->no_kamar . " tidak dapat dihapus!");
+        }
         return redirect('/kamar')->with('pesan', "kamar $kamar->nomor berhasil dihapus");
     }
 
     public function search(Request $request)
     {
         $kamars = Kamar::where('no_kamar', '=', $request->no_kamar)->get();
+        $jumlah = Kamar::where('no_kamar', '=', $request->no_kamar)->count();
+
+        if ($jumlah < 1) {
+            Alert::info("Tidak ada data :'(", "Kamar " . $request->no_kamar . " tidak ditemukan!");
+        }
         return view('kamar.index', compact('kamars'));
     }
 }
